aruha-563: moved check to separate class
