+++ src/test/java/org/zalando/nakadi/util/JsonUtilsTest.java
@@ -1,54 +1,50 @@
-package org.zalando.nakadi.service;
+package org.zalando.nakadi.util;
 
 import org.junit.Test;
 import org.zalando.nakadi.exceptions.InvalidEventTypeException;
 
-public class EventTypeServiceTest {
-
-    private final EventTypeService eventTypeService =
-            new EventTypeService(null, null, null, null, null, null, null, null);
+public class JsonUtilsTest {
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaEndingBracket() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("{\"additionalProperties\": true}}");
+        JsonUtils.checkEventTypeSchemaValid("{\"additionalProperties\": true}}");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaMultipleRoots() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("{\"additionalProperties\": true}{\"additionalProperties\": true}");
+        JsonUtils.checkEventTypeSchemaValid("{\"additionalProperties\": true}{\"additionalProperties\": true}");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaArbitraryEnding() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("{\"additionalProperties\": true}NakadiRocks");
+        JsonUtils.checkEventTypeSchemaValid("{\"additionalProperties\": true}NakadiRocks");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaArrayEnding() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("[{\"additionalProperties\": true}]]");
+        JsonUtils.checkEventTypeSchemaValid("[{\"additionalProperties\": true}]]");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaEndingCommaArray() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("[{\"test\": true},]");
+        JsonUtils.checkEventTypeSchemaValid("[{\"test\": true},]");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaEndingCommaArray2() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("[\"test\",]");
+        JsonUtils.checkEventTypeSchemaValid("[\"test\",]");
     }
 
     @Test(expected = InvalidEventTypeException.class)
     public void testValidateSchemaEndingCommaObject() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("{\"test\": true,}");
+        JsonUtils.checkEventTypeSchemaValid("{\"test\": true,}");
     }
 
     @Test
     public void testValidateSchemaFormattedJson() throws InvalidEventTypeException {
-        eventTypeService.checkJsonIsValid("{\"properties\":{\"event_class\":{\"type\":\"string\"},\"app_domain_id\":" +
+        JsonUtils.checkEventTypeSchemaValid("{\"properties\":{\"event_class\":{\"type\":\"string\"},\"app_domain_id\":" +
                 "{\"type\":\"integer\"},\"event_type\":{\"type\":\"string\"},\"time\":{\"type\":\"number\"}," +
                 "\"partitioning_key\":{\"type\":\"string\"},\"body\":{\"type\":\"object\"}}," +
                 "\"additionalProperties\":true}");
     }
-
 }
